% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cull_database_entries.R
\name{cull_database_entries}
\alias{cull_database_entries}
\title{cull_database_entries}
\usage{
cull_database_entries(table, outdir, file_name, con, machine_id = NULL)
}
\arguments{
\item{outdir}{Directory from which the file will be read, and where the delete_log_FILE_NAME will be read to}

\item{file_name}{The name of the file being read in}

\item{con}{connection the the bety database}

\item{machine_id}{Optional id of the machine that contains the bety entries.}
}
\description{
This is a fucntion that takes in a table of records and deletes everything in the file. Please do not run this function without 
1) Backing Up Bety 
2) Checking the the file only contains entries to be deleted. 

For more information on how to use this function see the "Pre-release-database-cleanup" script in the 'vignettes' folder
or look at the README
}
\author{
Tempest McCabe
}
