% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_formats_without_inputs.R
\name{find_formats_without_inputs}
\alias{find_formats_without_inputs}
\title{find_formats_without_inputs}
\usage{
find_formats_without_inputs(con, user_id = NULL, created_after = NULL,
  updated_after = NULL)
}
\arguments{
\item{con}{connection the the bety database}

\item{user_id}{Optional parameter to search by user_id}

\item{created_after}{Optional parameter to search by creation date. Date must be in form 'YYYY-MM-DD'}

\item{updated_after}{Optional parameter to search all entried updated after a certain date. Date must be in form 'YYYY-MM-DD'}
}
\description{
This is a fucntion that returns a dataframe with all of the format entries that have no assosiated input records.  

For more information on how to use this function see the "Pre-release-database-cleanup" script in the 'vignettes' folder
or look at the README
}
\author{
Tempest McCabe
}
