\name{graphOrthographic}
\alias{graphOrthographic}
\alias{graphSliceThickness}
\alias{graphSingleSlice}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Graphing Functions for QAQC Process}
\description{
  ...
}
\usage{
graphOrthographic(nim, uid, modality, imgPath, htmlFile, width = 750,
                  height = 750, cornerText = NULL, captionText = NULL)
graphSliceThickness(nim, uid, imagePositionPatient, movingDimensions,
                    imgPath, htmlFile, width = 500, height = 500)
graphSingleSlice(nim, uid, modality, imgPath, htmlFile, width = 750,
                 height = 750)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nim}{\ldots}
  \item{uid}{\ldots}
  \item{modality}{\ldots}
  \item{imagePositionPatient}{\ldots}
  \item{movingDimensions}{\ldots}
  \item{imgPath}{\ldots}
  \item{htmlFile}{\ldots}
  \item{width}{\ldots}
  \item{height}{\ldots}
  \item{cornerText}{\ldots}
  \item{captionText}{\ldots}
}
%\details{}
\value{
  A graph.
}
%\references{}
\author{Brandon Whitcher \email{bjw34032@users.sourceforge.net}}
%\note{}
%\seealso{}
%\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
