\name{new.taylor}
\alias{new.taylor}
\title{Taylor Diagram}
\usage{
  new.taylor(dataset, runid, siteid)
}
\arguments{
<<<<<<< HEAD
  \item{runid}{a numeric vector with the id(s) of one or
  more runs (folder in runs) to plot}
=======
  \item{dataset}{}

  \item{runid}{a numeric vector with the id(s) of one or
  more runs (folder in runs) to plot}

  \item{siteid}{}
>>>>>>> upstream/master
}
\description{
  plot taylor diagram for benchmark sites
}

