\name{parseDomain}
\alias{parseDomain}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Parse Domain Fields}
\description{
  A quick and (hopefully) efficient way of extracting specific
  information from Domain search output.
}
\usage{
parseDomain(dom.list, character.split = "/", index = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dom.list}{is the list of Domain search output.}
  \item{character.split}{is a character, or vector of characters, that
    divide the text string(s).}
  \item{index}{is the \dQuote{depth} of the splitting.}
}
\details{
  TBD.
}
\value{
  A character string containing the information from the desired Domain
  search output.
}
%\references{}
\author{Brandon Whitcher \email{bjw34032@gsk.com}}
%\note{
\seealso{\code{\link{qaqc}}}
%\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }

