% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_out_table.R
\name{write_out_table}
\alias{write_out_table}
\title{write_out_table}
\usage{
write_out_table(table, table_name, outdir, relevant_table_columns)
}
\arguments{
\item{table}{a table that is output from one of the find_* functions}

\item{table_name}{name of table}

\item{outdir}{path to folder into which the editable table will be written}

\item{relevant_table_columns}{a list of all columns to keep. ID and table name will be automatically included.}
}
\description{
This is a function that returns a dataframe with all of the format entries that have no associated input records.

For more information on how to use this function see the "Pre-release-database-cleanup" script in the 'vignettes' folder
or look at the README
}
\author{
Tempest McCabe
}
