% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cull_database_entries.R
\name{cull_database_entries}
\alias{cull_database_entries}
\title{Delete selected records from bety}
\usage{
cull_database_entries(
  table = NULL,
  outdir,
  file_name = NULL,
  con,
  machine_id = NULL,
  table_name = NULL
)
}
\arguments{
\item{table}{data frame containing records to be deleted. Specify either this or `file_name`}

\item{outdir}{Directory from which the file will be read, and where the delete_log_FILE_NAME will be read to}

\item{file_name}{The name of the file being read in. Specify either this or `table`}

\item{con}{connection the the bety database}

\item{machine_id}{Optional id of the machine that contains the bety entries.}

\item{table_name}{database table from which to delete}
}
\description{
This is a fucntion that takes in a table of records and deletes everything in the file. Please do not run this function without
1) Backing Up Bety
2) Checking the the file only contains entries to be deleted.

For more information on how to use this function see the "Pre-release-database-cleanup" script in the 'vignettes' folder
or look at the README
}
\author{
Tempest McCabe
}
