% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_formats_without_inputs.R
\name{find_formats_without_inputs}
\alias{find_formats_without_inputs}
\title{Find formats in bety that have no input record in bety}
\usage{
find_formats_without_inputs(
  con,
  user_id_code = NULL,
  created_after = NULL,
  updated_after = NULL,
  created_before = NULL,
  updated_before = NULL
)
}
\arguments{
\item{con}{connection the the bety database}

\item{user_id_code}{Optional parameter to search by user_id}

\item{created_after}{Optional parameter to search by creation date. Date must be in form 'YYYY-MM-DD'.}

\item{updated_after}{Optional parameter to search all entries updated after a certain date. Date must be in form 'YYYY-MM-DD'.}

\item{created_before}{Optional parameter to search by creation date. Can be used in conjunction with created_after to specify a specific window. Date must be in form 'YYYY-MM-DD'.}

\item{updated_before}{Optional parameter to search all entries updated before a certain date. Date must be in form 'YYYY-MM-DD'.}
}
\description{
This is a function that returns a dataframe with all of the format entries that have no associated input records.

For more information on how to use this function see the "Pre-release-database-cleanup" script in the 'vignettes' folder
or look at the README
}
\author{
Tempest McCabe
}
