% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_inputs_without_formats.R
\name{find_inputs_without_formats}
\alias{find_inputs_without_formats}
\title{Find inputs in bety with no format records}
\usage{
find_inputs_without_formats(
  con,
  user_id = NULL,
  created_after = NULL,
  updated_after = NULL,
  created_before = NULL,
  updated_before = NULL
)
}
\arguments{
\item{con}{connection the the bety database}

\item{user_id}{Optional parameter to search by user_id}

\item{created_before, created_after}{Optional parameter to search by creation date. Date must be in form 'YYYY-MM-DD'}

\item{updated_before, updated_after}{Optional parameter to search all entried updated after a certain date. Date must be in form 'YYYY-MM-DD'}
}
\description{
This is a function that returns a dataframe with all of the input entries that have no associated format records.
This is very rare in the database.

For more information on how to use this function see the "Pre-release-database-cleanup" script in the 'vignettes' folder
or look at the README
}
\author{
Tempest McCabe
}
