\name{new.taylor}
\alias{new.taylor}
\title{Taylor Diagram}
\usage{
  new.taylor(dataset, runid, siteid)
}
\arguments{
<<<<<<< HEAD
  \item{dataset}{}

  \item{runid}{a numeric vector with the id(s) of one or
  more runs (folder in runs) to plot}

  \item{siteid}{}
=======
  \item{runid}{a numeric vector with the id(s) of one or
  more runs (folder in runs) to plot}
>>>>>>> ef9572c0bfd74f64f90e160d3d29f5e3f933b0b6
}
\description{
  plot taylor diagram for benchmark sites
}

