% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor.plot.R
\name{new.taylor}
\alias{new.taylor}
\title{Plot taylor diagram for benchmark sites}
\usage{
new.taylor(dataset, runid, siteid)
}
\arguments{
\item{dataset}{data to plot}

\item{runid}{a numeric vector with the id(s) of one or more runs (folder in runs) to plot}

\item{siteid}{vector of sites to plot}
}
\description{
Plot taylor diagram for benchmark sites
}
