% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor.plot.R
\name{new.taylor}
\alias{new.taylor}
\title{Taylor Diagram}
\usage{
new.taylor(dataset, runid, siteid)
}
\arguments{
\item{runid}{a numeric vector with the id(s) of one or more runs (folder in runs) to plot}
}
\description{
Plot taylor diagram for benchmark sites
}
